requirements.txt
discord.py>=2.3.2
asyncpg>=0.29.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
pandas>=2.2.0
tabulate>=0.9.0
opgg.py>=3.1.0
